+++ copilot+translate/edge+chrome/background.js
@@ -1,40 +1,27 @@
-const BING_ORIGIN = 'https://www.bing.com';
+// Allows users to open the side panel by clicking on the action toolbar icon
+chrome.sidePanel
+  .setPanelBehavior({openPanelOnActionClick: true})
+  .catch((error) => console.error(error))
 
 chrome.runtime.onInstalled.addListener(() => {
-  // chrome.contextMenus.create({
-  //   id: 'openSidePanel',
-  //   title: 'Open sidebar',
-  //   contexts: ['all']
-  // });
-});
+  // 在右鍵選單中新增 iCloud Copilot 選項
+  chrome.contextMenus.create({
+    id: 'icloud-copilot',
+    title: `${chrome.i18n.getMessage('extensionName')}`,
+    contexts: ['all'],
+  })
+})
 
-// chrome.contextMenus.onClicked.addListener((info, tab) => {
-//   if (info.menuItemId === 'openSidePanel') {
-//     // Open the sidebar in all the tabs of the current window.
-//     chrome.sidePanel.open({ windowId: tab.windowId });
-//   }
-// });
-
-chrome.tabs.onUpdated.addListener(async (tabId, info, tab) => {
-  if (!tab.url) return;
-  const url = new URL(tab.url);
-  // Enables the sidebar when at bing.com
-  if (url.origin === BING_ORIGIN) {
-    await chrome.sidePanel.setOptions({
-      tabId,
-      path: 'sidepanel.html',
-      enabled: true
-    });
-  } else {
-    // Disables the sidebar when at other sites
-    await chrome.sidePanel.setOptions({
-      tabId,
-      enabled: false
-    });
+// 當使用者點擊右鍵選單中的 iCloud Copilot 選項時，開啟側邊欄
+chrome.contextMenus.onClicked.addListener(async (info, tab) => {
+  if (info.menuItemId === 'icloud-copilot') {
+    // This will open the panel in all the pages on the current window.
+    await chrome.sidePanel.open({windowId: tab.windowId})
+    // wait for the panel to be opened
+    await new Promise((resolve) => setTimeout(resolve, 200))
+    chrome.runtime.sendMessage({
+      name: info.menuItemId,
+      data: info,
+    })
   }
-});
-
-// Allow users to open the sidebar by clicking the action toolbar icon
-chrome.sidePanel
-  .setPanelBehavior({ openPanelOnActionClick: true })
-  .catch((error) => console.error(error));
+})