+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 
 ### 單一微服務(容器所代表的服務)
 - **:whale2:Docker**：服務建置工具。
-- **Translate**：翻譯擴充套件。
+- **Copilot&Translate**：聊天機器人暨翻譯擴充套件。
 
 ## 版本
 <!-- no toc -->